base: fixed warning in DataVector serialized vector parser datadriven(OCL): - changed pattern of AutoTuningPaper experiment result file names

+++ base/src/sgpp/base/datatypes/DataVector.cpp
@@ -96,20 +96,20 @@ DataVector DataVector::fromString(const std::string& serializedVector) {
       i += 1;
     } else if (state == PARSER_STATE::VALUE) {
       //      size_t next;
-//      double value = std::atof(&(serializedVector[i]));
-    	int endNumber = i;
-        while (serializedVector[endNumber] != ',' && serializedVector[endNumber] != ']') endNumber++;
-    	std::stringstream stream;
-    	for (size_t j = i; j < endNumber; j++) {
-    		stream << serializedVector[j];
-    	}
-    	std::string shortString(stream.str());
-//      double value = std::atof(shortString.c_str());
-    	double value = std::stod(shortString);
+      //      double value = std::atof(&(serializedVector[i]));
+      size_t endNumber = i;
+      while (serializedVector[endNumber] != ',' && serializedVector[endNumber] != ']') endNumber++;
+      std::stringstream stream;
+      for (size_t j = i; j < endNumber; j++) {
+        stream << serializedVector[j];
+      }
+      std::string shortString(stream.str());
+      //      double value = std::atof(shortString.c_str());
+      double value = std::stod(shortString);
       v.append(value);
       state = PARSER_STATE::COMMAEND;
       //      i += next;
-//      while (serializedVector[i] != ',' && serializedVector[i] != ']') i++;
+      //      while (serializedVector[i] != ',' && serializedVector[i] != ']') i++;
       i = endNumber;
 
     } else if (state == PARSER_STATE::COMMAEND) {
